+++ api/src/main/java/api/domain/fcm/converter/FcmTokenConverter.java
@@ -1,6 +1,6 @@
 package api.domain.fcm.converter;
 
-import api.domain.admin.controller.model.response.FcmTokenInfoList;
+import api.domain.fcm.controller.model.FcmTokenInfo;
 import api.domain.fcm.controller.model.FcmTokenRequest;
 import db.domain.token.fcm.FcmTokenDocument;
 import global.annotation.Converter;
@@ -17,8 +17,8 @@ public class FcmTokenConverter {
             .build();
     }
 
-    public FcmTokenInfoList toResponse(FcmTokenDocument fcmTokenDocument) {
-        return FcmTokenInfoList.builder()
+    public FcmTokenInfo toResponse(FcmTokenDocument fcmTokenDocument) {
+        return FcmTokenInfo.builder()
             .fcmToken(fcmTokenDocument.getFcmToken())
             .generalNewsTopic(fcmTokenDocument.isGeneralNewsTopic())
             .scholarshipNewsTopic(fcmTokenDocument.isScholarshipNewsTopic())
@@ -30,7 +30,7 @@ public class FcmTokenConverter {
             .build();
     }
 
-    public List<FcmTokenInfoList> toListResponse(List<FcmTokenDocument> fcmTokenDocumentList) {
+    public List<FcmTokenInfo> toListResponse(List<FcmTokenDocument> fcmTokenDocumentList) {
         return fcmTokenDocumentList.stream()
             .map(this::toResponse)
             .toList();
