+++ api/src/main/java/api/domain/notice/controller/model/request/NoticeSearchRequest.java
@@ -1,6 +1,5 @@
-package api.domain.notice.controller.model.search;
+package api.domain.notice.controller.model.request;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
