+++ api/src/main/java/api/domain/report/controller/model/ReportSimpleResponse.java
@@ -1,4 +1,4 @@
-package api.domain.admin.controller.model.response;
+package api.domain.report.controller.model;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 @Builder
-public class ReportListResponse {
+public class ReportSimpleResponse {
 
     private String reportId;
