+++ api/src/main/java/api/domain/urgent/controller/model/UrgentNoticeSaveRequest.java
@@ -1,4 +1,4 @@
-package api.domain.admin.controller.model.request;
+package api.domain.urgent.controller.model;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
